requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
paho-mqtt==1.6.1
requests==2.31.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
